# perf(GQL): RHICOMPL-3218 batch-load rule-result related fields