--- a/app/graphql/types/concerns/test_results.rb
+++ b/app/graphql/types/concerns/test_results.rb
@@ -28,28 +28,33 @@ module Types
 
       def compliant(args = {})
         latest_test_result_batch(args).then do |latest_test_result|
-          host_results = latest_test_result&.rule_results
-          host_results.present? &&
-            latest_test_result.score >= object.compliance_threshold
-        end
-      end
-
-      def rules_passed(args = {})
-        latest_test_result_batch(args).then do |latest_test_result|
           if latest_test_result.blank?
-            0
+            false
           else
-            latest_test_result.rule_results.passed.count
+            CollectionLoader.for(::TestResult, :has_rule_results).load(latest_test_result).then do |has|
+              has.present? &&
+                latest_test_result.score >= object.compliance_threshold
+            end
           end
         end
       end
 
+      def rules_passed(args = {})
+        rules_count(args, RuleResult::PASSED)
+      end
+
       def rules_failed(args = {})
+        rules_count(args, RuleResult::FAIL)
+      end
+
+      def rules_count(args, subset)
         latest_test_result_batch(args).then do |latest_test_result|
           if latest_test_result.blank?
             0
           else
-            latest_test_result.rule_results.failed.count
+            CollectionLoader.for(::TestResult, :counters).load(latest_test_result).then do |counters|
+              counters.select { |counter| subset.include?(counter.result) }.sum { |item| item&.count.to_i }
+            end
           end
         end
       end