--- a/app/models/test_result.rb
+++ b/app/models/test_result.rb
@@ -11,6 +11,13 @@ class TestResult < ApplicationRecord
   has_many :failed_rule_results, -> { failed }, inverse_of: :test_result,
                                                 class_name: 'RuleResult',
                                                 dependent: :delete_all
+  has_many :counters, lambda {
+    group(:result, :test_result_id).select(:result, 'COUNT(result)', :test_result_id)
+  }, class_name: 'RuleResult', dependent: :nullify, inverse_of: :test_result
+
+  has_one :has_rule_results, lambda {
+    select(1, :test_result_id)
+  }, class_name: 'RuleResult', dependent: :nullify, inverse_of: :test_result
 
   validates :host, presence: true, on: :create
   validates :host_id, presence: true,