Tried the following query with 600 systems:
```
{
  "operationName": "supportedProfilesByOSMajor",
  "variables": { "perPage": 500,  "page": 1, "tagsColumn": null, "filter": "(policy_id = f9d3fe39-e586-4ef8-a863-deea6e70577d)",  "policyId": "f9d3fe39-e586-4ef8-a863-deea6e70577d" },
  "query": "
  query getSystems(
    $filter: String!
    $policyId: ID
    $perPage: Int
    $page: Int
    $sortBy: [String!]
    $tags: [String!]
  ) {
    systems(
      search: $filter
      limit: $perPage
      offset: $page
      sortBy: $sortBy
      tags: $tags
    ) {
      totalCount
      edges {
        node {
          id
          name
          osMajorVersion
          osMinorVersion
          insightsId
          testResultProfiles(policyId: $policyId) {
            lastScanned
            compliant
            score
            supported

            benchmark {
              version
            }
            rulesFailed
          }
        }
      }
    }
  }"
}
```

Before:
```
Completed 200 OK in 7270ms (Views: 3.6ms | ActiveRecord: 484.2ms | Allocations: 6130762)
```
After:
```
Completed 200 OK in 2058ms (Views: 5.2ms | ActiveRecord: 334.5ms | Allocations: 1301295)
```

## Secure Coding Practices Checklist GitHub Link
- https://github.com/RedHatInsights/secure-coding-checklist

## Secure Coding Checklist
- [ ] Input Validation
- [ ] Output Encoding
- [ ] Authentication and Password Management
- [ ] Session Management
- [ ] Access Control
- [ ] Cryptographic Practices
- [ ] Error Handling and Logging
- [ ] Data Protection
- [ ] Communication Security
- [ ] System Configuration
- [ ] Database Security
- [ ] File Management
- [ ] Memory Management
- [x] General Coding Practices